# internal/docs/swagger.yaml
definitions:
  github_com_umar5678_go-backend_internal_models.TransactionStatus:
    enum:
    - pending
    - completed
    - failed
    - cancelled
    - held
    - released
    type: string
    x-enum-varnames:
    - TransactionStatusPending
    - TransactionStatusCompleted
    - TransactionStatusFailed
    - TransactionStatusCancelled
    - TransactionStatusHeld
    - TransactionStatusReleased
  github_com_umar5678_go-backend_internal_models.TransactionType:
    enum:
    - credit
    - debit
    - refund
    - hold
    - release
    - transfer
    type: string
    x-enum-varnames:
    - TransactionTypeCredit
    - TransactionTypeDebit
    - TransactionTypeRefund
    - TransactionTypeHold
    - TransactionTypeRelease
    - TransactionTypeTransfer
  github_com_umar5678_go-backend_internal_models.UserRole:
    enum:
    - rider
    - driver
    - admin
    - delivery_person
    - service_provider
    - handyman
    type: string
    x-enum-varnames:
    - RoleRider
    - RoleDriver
    - RoleAdmin
    - RoleDeliveryPerson
    - RoleServiceProvider
    - RoleHandyman
  github_com_umar5678_go-backend_internal_models.UserStatus:
    enum:
    - active
    - suspended
    - banned
    - pending_verification
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusSuspended
    - StatusBanned
    - StatusPendingVerification
  github_com_umar5678_go-backend_internal_models.WalletType:
    enum:
    - rider
    - driver
    - platform
    type: string
    x-enum-varnames:
    - WalletTypeRider
    - WalletTypeDriver
    - WalletTypePlatform
...
  github_com_umar5678_go-backend_internal_modules_auth_dto.UserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: string
      lastLoginAt:
        type: string
      name:
        type: string
      phone:
        type: string
      profilePhotoUrl:
        type: string
      role:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_models.UserRole'
      status:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_models.UserStatus'
    type: object
...
  github_com_umar5678_go-backend_internal_modules_wallet_dto.AddFundsRequest:
    properties:
      amount:
        type: number
      description:
        type: string
    required:
    - amount
    type: object
  github_com_umar5678_go-backend_internal_modules_wallet_dto.CaptureHoldRequest:
    properties:
      amount:
        description: 'Optional: capture partial amount'
        type: number
      description:
        type: string
      holdId:
        type: string
    required:
    - holdId
    type: object
  github_com_umar5678_go-backend_internal_modules_wallet_dto.HoldFundsRequest:
    properties:
      amount:
        type: number
      holdDuration:
        description: minutes, default 30
        minimum: 1
        type: integer
      referenceId:
        type: string
      referenceType:
        type: string
    required:
    - amount
    - referenceId
    - referenceType
    type: object
  github_com_umar5678_go-backend_internal_modules_wallet_dto.HoldResponse:
    properties:
      amount:
        type: number
      createdAt:
        type: string
      expiresAt:
        type: string
      id:
        type: string
      referenceId:
        type: string
      referenceType:
        type: string
      releasedAt:
        type: string
      status:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_models.TransactionStatus'
      walletId:
        type: string
    type: object
  github_com_umar5678_go-backend_internal_modules_wallet_dto.ReleaseHoldRequest:
    properties:
      holdId:
        type: string
    required:
    - holdId
    type: object
  github_com_umar5678_go-backend_internal_modules_wallet_dto.TransactionResponse:
    properties:
      amount:
        type: number
      balanceAfter:
        type: number
      balanceBefore:
        type: number
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      metadata:
        additionalProperties: true
        type: object
      processedAt:
        type: string
      referenceId:
        type: string
      referenceType:
        type: string
      status:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_models.TransactionStatus'
      type:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_models.TransactionType'
      walletId:
        type: string
    type: object
  github_com_umar5678_go-backend_internal_modules_wallet_dto.TransferFundsRequest:
    properties:
      amount:
        type: number
      description:
        type: string
      recipientId:
        type: string
    required:
    - amount
    - recipientId
    type: object
  github_com_umar5678_go-backend_internal_modules_wallet_dto.WalletResponse:
    properties:
      availableBalance:
        type: number
      balance:
        type: number
      createdAt:
        type: string
      currency:
        type: string
      heldBalance:
        type: number
      id:
        type: string
      isActive:
        type: boolean
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_auth_dto.UserResponse'
      userId:
        type: string
      walletType:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_models.WalletType'
    type: object
  github_com_umar5678_go-backend_internal_modules_wallet_dto.WithdrawFundsRequest:
    properties:
      amount:
        type: number
      description:
        type: string
    required:
    - amount
    type: object
  github_com_umar5678_go-backend_internal_utils_response.ErrorDetail:
    properties:
      code:
        type: string
      field:
        type: string
      message:
        type: string
    type: object
  github_com_umar5678_go-backend_internal_utils_response.Meta:
    properties:
      durationMs:
        type: number
      method:
        type: string
      path:
        type: string
      requestId:
        type: string
      timestamp:
        type: string
      version:
        type: string
    type: object
  github_com_umar5678_go-backend_internal_utils_response.Response:
    properties:
      code:
        type: string
      data: {}
      errors:
        description: Use ErrorDetail
        items:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.ErrorDetail'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Meta'
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: Production-ready Go backend API
  title: supr booking server in go
  version: "1.0"
paths:
...
  /wallet:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_wallet_dto.WalletResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get wallet details
      tags:
      - wallet
  /wallet/add-funds:
    post:
      consumes:
      - application/json
      parameters:
      - description: Add funds data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_wallet_dto.AddFundsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_wallet_dto.TransactionResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Add funds to wallet
      tags:
      - wallet
  /wallet/balance:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  format: float64
                  type: number
              type: object
      security:
      - BearerAuth: []
      summary: Get wallet balance
      tags:
      - wallet
  /wallet/hold:
    post:
      consumes:
      - application/json
      parameters:
      - description: Hold funds data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_wallet_dto.HoldFundsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_wallet_dto.HoldResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Place a hold on funds
      tags:
      - wallet
  /wallet/hold/capture:
    post:
      consumes:
      - application/json
      parameters:
      - description: Capture hold data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_wallet_dto.CaptureHoldRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_wallet_dto.TransactionResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Capture a hold
      tags:
      - wallet
  /wallet/hold/release:
    post:
      consumes:
      - application/json
      parameters:
      - description: Hold ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_wallet_dto.ReleaseHoldRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
      security:
      - BearerAuth: []
      summary: Release a hold
      tags:
      - wallet
  /wallet/transactions:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Transaction type
        in: query
        name: type
        type: string
      - description: Transaction status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_wallet_dto.TransactionResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List wallet transactions
      tags:
      - wallet
  /wallet/transactions/{id}:
    get:
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_wallet_dto.TransactionResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Get transaction details
      tags:
      - wallet
  /wallet/transfer:
    post:
      consumes:
      - application/json
      parameters:
      - description: Transfer data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_wallet_dto.TransferFundsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_wallet_dto.TransactionResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Transfer funds to another user
      tags:
      - wallet
  /wallet/withdraw:
    post:
      consumes:
      - application/json
      parameters:
      - description: Withdraw funds data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_wallet_dto.WithdrawFundsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_umar5678_go-backend_internal_utils_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_umar5678_go-backend_internal_modules_wallet_dto.TransactionResponse'
              type: object
      security:
      - BearerAuth: []
      summary: Withdraw funds from wallet
      tags:
      - wallet
swagger: "2.0"